# Speed up polynomial computation while improve Chunk speed a bit