Hi, I've recently referred to this pkg and related papers to learn about CDC, and in the process I've made some optimizations for polynomial computation. I also did some minor optimizations for Chunk in the process, which probably had a ~3% speedup:
```plain
goos: linux
goarch: amd64
pkg: github.com/restic/chunker
cpu: Intel(R) Xeon(R) Platinum 8269CY CPU @ 2.50GHz
                    │    old.txt    │                new.txt                 │
                    │    sec/op     │    sec/op      vs base                 │
ChunkerWithSHA256-8    188.1m ±  0%    184.5m ±  0%    -1.89% (p=0.000 n=10)
Chunker-8              77.78m ±  0%    74.29m ±  0%    -4.49% (p=0.000 n=10)
NewChunker-8          102.31µ ± 23%    99.77µ ± 21%         ~ (p=0.353 n=10)
PolMul-8               195.9n ±  0%    127.5n ±  0%   -34.87% (p=0.000 n=10)
PolDiv-8               86.88n ±  0%    86.79n ±  0%    -0.10% (p=0.049 n=10)
PolDeg-8              0.3744n ±  0%   0.7498n ±  0%  +100.29% (p=0.000 n=10)
PolMod-8               86.89n ±  0%    86.80n ±  0%         ~ (p=0.578 n=10)
PolDivMod-8            86.29n ±  0%    86.30n ±  0%         ~ (p=0.540 n=10)
RandomPolynomial-8    2471.3µ ±  3%    377.7µ ±  1%   -84.72% (p=0.000 n=10)
PolIrreducible-8      1372.1µ ±  0%    130.1µ ±  0%   -90.52% (p=0.000 n=10)
PolGCD-8               789.5n ±  0%    738.2n ±  0%    -6.50% (p=0.000 n=10)
PolMulMod-8           9995.0n ±  0%    555.0n ±  0%   -94.45% (p=0.000 n=10)
geomean                10.56µ          5.881µ         -44.28%

                    │   old.txt    │               new.txt               │
                    │     B/s      │     B/s       vs base               │
ChunkerWithSHA256-8   170.1Mi ± 0%   173.4Mi ± 0%  +1.93% (p=0.000 n=10)
Chunker-8             411.4Mi ± 0%   430.8Mi ± 0%  +4.70% (p=0.000 n=10)
geomean               264.6Mi        273.3Mi       +3.30%
```

These are my test results in my homlab, it may not be accurate, if you have time please double check it. Meanwhile, I tested compiling restic with this optimized version and the program works fine(🤔️).