@wjiec 2023-09-12
inline `slide` into `reset` for greater performance gains:
```plain
goos: linux
goarch: amd64
pkg: github.com/restic/chunker
cpu: Intel(R) Xeon(R) Platinum 8163 CPU @ 2.50GHz
                    │    old.txt    │                 new.txt                 │
                    │    sec/op     │    sec/op      vs base                  │
ChunkerWithSHA256-8    189.3m ±  1%    184.3m ±  0%     -2.62% (p=0.000 n=10)
Chunker-8              79.32m ±  2%    74.19m ±  1%     -6.47% (p=0.000 n=10)
NewChunker-8           214.5µ ± 10%    125.3µ ± 46%    -41.55% (p=0.000 n=10)
PolDivMod-8            3.463n ±  1%   96.620n ±  0%  +2690.47% (p=0.000 n=10)
PolDiv-8               3.463n ±  1%   96.385n ±  0%  +2683.68% (p=0.000 n=10)
PolMod-8               3.463n ±  1%   96.380n ±  0%  +2682.73% (p=0.000 n=10)
PolDeg-8              0.3743n ±  0%   0.3748n ±  0%          ~ (p=0.382 n=10)
RandomPolynomial-8    2436.3µ ±  2%    363.6µ ±  2%    -85.08% (p=0.000 n=10)
PolIrreducible-8      1363.0µ ±  0%    127.6µ ±  0%    -90.64% (p=0.000 n=10)
PolMul-8                               121.5n ±  0%
PolGCD-8                               712.6n ±  0%
PolMulMod-8                            557.2n ±  0%
geomean                8.192µ          5.744µ          +75.87%

                    │   old.txt    │               new.txt               │
                    │     B/s      │     B/s       vs base               │
ChunkerWithSHA256-8   169.1Mi ± 1%   173.6Mi ± 0%  +2.69% (p=0.000 n=10)
Chunker-8             403.4Mi ± 2%   431.3Mi ± 1%  +6.92% (p=0.000 n=10)
geomean               261.2Mi        273.7Mi       +4.78%
```