@MichaelEischer 2023-09-27
Please split the polynomial-related changes into a separate PR. I'm currently pretty short of time, and I'd prefer to handle the changes in `chunker.go` first, as these are performance critical for restic in contrast to the polynomial changes.